pyproject.toml
[project]
name = "mate-2026-blue-lobster"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dearpygui>=2.1.1",
    "isort>=7.0.0",
    "nicegui>=3.5.0",
    "numpy>=2.2.6",
    "opencv-python>=4.12.0.88",
    "pandas>=2.3.3",
    "pycairo>=1.29.0",
    "pygobject>=3.54.5",
    "pywebview>=6.1",
    "ruff>=0.14.4",
    "ultralytics>=8.3.228",
]
